// api/types.ts
export interface UrlItem {
  url: string;
  lastUsed?: string;
}

export interface AccountItem {
  id: number;
  username: string;
  password: string;
  relatedIds: number[];
  createdAt: string;
  lastUsed: string;
}

export interface Bookmark {
  id: string;
  name: string;
  pinyinInitials: string;
  tagIds: number[];
  urls: UrlItem[];
  notes: string;
  accounts: AccountItem[];
  createdAt: string;
  updatedAt: string;
  lastUsedAt: string;
}

export interface BookmarkCreate {
  name: string;
  pinyinInitials?: string;
  tagIds?: number[];
  urls?: UrlItem[];
  notes?: string;
  accounts?: AccountCreate[];
}

export interface AccountCreate {
  username: string;
  password: string;
  relatedIds?: number[];
}